requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Core
pydantic>=2.0
python-dotenv>=1.0.0

# ML / NLP
torch>=2.0.0
transformers>=4.40.0
pandas>=2.0.0

# HTTP & DB
httpx>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0

# Google Cloud
google-cloud-texttospeech>=2.14.0
google-cloud-speech>=2.25.0

# Google ADK & GenAI (Live API agent)
google-adk>=1.25.0
google-genai>=1.0.0
accelerate
